fix: align Player segment ID type with Engine

+++ app/src/main/java/com/looplingua/app/data/looplingua/PlayerProjectMapper.kt
@@ -31,7 +31,7 @@ class PlayerProjectMapper {
                 loopLinguaTrack.segments.map { loopLinguaSegment ->
 
                     Segment(
-                        id = loopLinguaSegment.segmentId.toLong(),
+                        id = loopLinguaSegment.segmentId,
 
                         originalStartMs =
                             loopLinguaSegment.originalStartMs,
